@@ -1,10 +1,8 @@
-﻿using Bitrix24RestApiClient.Core.Builders.Interfaces;
-using Bitrix24RestApiClient.Core.Models.Enums;
+﻿using Bitrix24RestApiClient.Core.Models.Enums;
 
-namespace Bitrix24RestApiClient.Core.Builders.Interfaces
+namespace Bitrix24RestApiClient.Core.Builders.Interfaces;
+
+public interface IEmailListBuilder
 {
-    public interface IEmailListBuilder
-    {
-        IEmailListBuilder SetField(string email, string type = EmailType.Рабочий);
-    }
+    IEmailListBuilder SetField(string email, string type = EmailType.Рабочий);
 }
