@@ -1,10 +1,8 @@
-﻿using Bitrix24RestApiClient.Core.Builders.Interfaces;
-using Bitrix24RestApiClient.Core.Models.Enums;
+﻿using Bitrix24RestApiClient.Core.Models.Enums;
 
-namespace Bitrix24RestApiClient.Core.Builders.Interfaces
+namespace Bitrix24RestApiClient.Core.Builders.Interfaces;
+
+public interface IPhoneListBuilder
 {
-    public interface IPhoneListBuilder
-    {
-        IPhoneListBuilder SetField(string phone, string type = EmailType.Рабочий);
-    }
+    IPhoneListBuilder SetField(string phone, string type = EmailType.Рабочий);
 }
