@@ -1,16 +1,14 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
 using Bitrix24RestApiClient.Core.Models.Response.Common;
-using Bitrix24RestApiClient.Core.Models.Response.FieldsResponse;
 
-namespace Bitrix24RestApiClient.Core.Models.Response.FieldsResponse
+namespace Bitrix24RestApiClient.Core.Models.Response.FieldsResponse;
+
+public class FieldsResponse
 {
-    public class FieldsResponse
-    {
-        [JsonProperty("result")]
-        public Dictionary<string, FieldInfo> Result { get; set; } = new Dictionary<string, FieldInfo>();
+    [JsonProperty("result")]
+    public Dictionary<string, FieldInfo> Result { get; set; } = new();
 
-        [JsonProperty("time")]
-        public Time Time { get; set; }
-    }
+    [JsonProperty("time")]
+    public Time Time { get; set; }
 }
