@@ -1,5 +1,4 @@
-﻿using System;
-using Bitrix24RestApiClient.Core.Models.Enums;
+﻿using Bitrix24RestApiClient.Core.Models.Enums;
 
 namespace Bitrix24RestApiClient.Core.Attributes;
 
