@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Bitrix24RestApiClient.Core.Models.Enums;
+﻿using Bitrix24RestApiClient.Core.Models.Enums;
 using Bitrix24RestApiClient.Core.Builders.Interfaces;
 using Bitrix24RestApiClient.Core.Models.CrmMultiField;
 using Bitrix24RestApiClient.Core.Models.CrmMultiField.implementations;
@@ -16,8 +15,5 @@ public class EmailListBuilder: IEmailListBuilder
         return this;
     }
 
-    public List<CrmMultiField> Build()
-    {
-        return fields;
-    }
+    public List<CrmMultiField> Build() => fields;
 }
